requirements.txt
Flask==3.1.0
Flask-Assets==2.1.0
joblib==1.4.2
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
Pillow==10.2.0
poetry==2.1.2
scikit-learn==1.5.0
scikit-image==0.22.0
seaborn==0.13.2
tensorflow==2.19.0
tqdm==4.67.1
Werkzeug==3.1.3
